pyproject.toml
[tool.poetry]
name = "ticket-triage-ml"
version = "0.1.0"
description = "Автоматическая классификация и приоритизация тикетов техподдержки с использованием многозадачного обучения"
authors = ["ML Team <ml@example.com>"]
readme = "README.md"
packages = [{include = "ticket_triage_ml"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "<2.0.0"
torch = "^2.1.0"
pytorch-lightning = "^2.1.0"
transformers = "^4.35.0"
datasets = "^2.15.0"
pandas = "^2.1.0"
pyarrow = "^14.0.0"
scikit-learn = "^1.3.0"
onnx = "^1.15.0"
onnxruntime = "^1.16.0"
fire = "^0.5.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
mlflow = "^2.9.0"
pydantic = "^2.5.0"
loguru = "^0.7.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
requests = "^2.31.0"
tqdm = "^4.66.0"
dvc = {version = "^3.30.0", extras = ["s3"]}
fastapi = "^0.109.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
rich = "^14.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
ticket-triage = "ticket_triage_ml.commands:main"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["ticket_triage_ml"]
skip_gitignore = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
